requirements.txt
# Image Compressor - Application Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
Pillow>=10.0.0

# Video compression
ffmpeg-python>=0.2.0

# Optional: for better performance
# numpy>=1.24.0
